requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
# alembic # Опционально, для миграций БД
